+++ src/CRON.go
@@ -126,6 +126,7 @@ func CRON() {
 		s.Every(1).Day().At("00:00").Do(func() {
 			utils.CleanupByDate("notifications")
 			utils.CleanupByDate("events")
+			imageCleanUp()
 		})
 		s.Start()
 	}()